Feature/#48 중간점검

+++ src/components/databox/DBBox.tsx
@@ -1,26 +1,28 @@
 import {FoodDto} from "../../interface/DataInterface";
-import {DBAtom, EditMode} from "../../recoil/DBAtom";
+import {DBAtom, EditMode, isLoginModalOpen} from "../../recoil/DBAtom";
 import {useRecoilState, useRecoilValue} from "recoil";
 import {EditOff} from "./isEdit/editOff";
 import {useEffect} from "react";
-import {useAxiosInstance} from "../../Axios/api";
+import {checkTokenValidate, useAxiosInstance} from "../../Axios/api";
 
 export const DBBox = ({ setAdd }: { setAdd: Function }) => {
   const [data, setData] = useRecoilState<FoodDto[]>(DBAtom);
-  const edit = useRecoilValue(EditMode);
+  const [, setIsLoginModalOpen] = useRecoilState(isLoginModalOpen)
+
   const instance = useAxiosInstance();
 
   useEffect(() => {}, [data]);
 
-  const handleRemoveFood = async (id: number) => {
+  const handleRemoveFood = async (id: number, name:string) => {
     try {
+      if (!window.confirm(`${name}을(를) 삭제 하시겠습니까?`)) return
       const response = await instance.delete(`food/${id}/recipes`);
       const newData = [...data];
       setData(newData.filter((item) => item.id !== id));
-
       alert("삭제되었습니다.");
-    } catch (err) {
-      console.log(err);
+    } catch (error) {
+      checkTokenValidate(error, setIsLoginModalOpen)
+      console.log(error);
     }
   };
   const buttonDesign = () => {
